map/requirements.txt
requests==2.26.0
grpcio==1.42.0
grpcio-tools==1.42.0
python-dotenv==0.19.2
